Tests: move browsers.js to flags folder
Ref gh-508
